writeFile in ShelterTrackerApplication.java updated to create writable file on first launch of program. test_file.json will save but no other json or xml files are persistent.

+++ ICS372_AndroidStart-master/ShelterTracker/app/src/main/java/edu/metrostate/sheltertracker/ShelterTrackerApplication.java
@@ -15,51 +15,32 @@ public class ShelterTrackerApplication extends Application {
 
     private final ShelterList shelterList = new ShelterList();
 
-
     @Override
     public void onCreate() {
         super.onCreate();
         String localFileLocation = getExternalFilesDir(null).getAbsolutePath()+File.separator+"seedFile.json";
         shelterList.addHashMap(ParseUtilities.loadJSON(localFileLocation));
-
-//        writeFile();
+        writeFile();
     }
 
     public ShelterList getShelterList() {
         return shelterList;
     }
 
-
-
     public void writeFile() {
-
         // this will put files in the /sdcard/Android/data/edu.metrostate.sheltertracker/files directory
         File externalDir = getExternalFilesDir(null);
-
         File outputFile = new File(externalDir, "seedFile.json");
-
-        try {
-            Files.createFile(outputFile.toPath());
-            Files.write(outputFile.toPath(), "My data".getBytes());
-
-        } catch (IOException ex) {
-            Log.e("FileCreation", "Error creating file", ex);
-        }
-
+        String localFileLocation = getExternalFilesDir(null).getAbsolutePath()+File.separator+"seedFile.json";
+        File file = new File(localFileLocation);
+        if(!file.exists()) {
+            try {
+                Files.createFile(outputFile.toPath());
+                Files.write(outputFile.toPath(), "My data".getBytes());
+            } catch (IOException ex) {
+                Log.e("FileCreation", "Error creating file", ex);
+            }
     }
 
-//    public String readExternalFile(String fileName) throws IOException {
-//        File path = getExternalFilesDir(null);
-//        File file = new File(path, fileName);
-//        List <String> temp_list = Files.readAllLines(file.toPath());
-//        return String.join("", temp_list);
-
-
-
-//        File path = getExternalFilesDir(null);
-//        List <String> temp_list = Files.readAllLines(path.toPath());
-//        String joined_string = String.join("", temp_list);
-//        ParseUtilities.addIncomingJSON(joined_string, shelterList);
-
-//    }
+    }
 }
